+++ libmatekbd/matekbd-keyboard-config.h
@@ -40,7 +40,7 @@ struct _MatekbdKeyboardConfig {
 
 	/* private, transient */
 	GSettings *settings;
-	int config_listener_id;
+	gulong config_listener_id;
 	XklEngine *engine;
 };
